my-project-hijack EFM32TG V0.01 1. NEW: achieve decode part. 2. CHG: test result send: 19792B, 3298 strings; rev: 3059 strings; PER = 1 - 3059/3298 = 7.25%.

+++  my-project-hihack --username kaiser.ren@gmail.com/energymicro/kits/EFM32TG_STK3300/examples/an0054_efm32_phone_audio_jack_interface/tg840/hijack_decode.h
@@ -71,8 +71,8 @@ typedef enum
  *        External Variable
  *----------------------------------------------------------------------------*/
 extern bool	edge_occur;
-extern uint16_t cur_stamp ;
-extern uint16_t prv_stamp ;
+extern uint32_t cur_stamp ;
+extern uint32_t prv_stamp ;
 extern edge_t 	cur_edge;
 
 /*----------------------------------------------------------------------------
